+++ ui/package.json
@@ -11,12 +11,14 @@
     "@mui/material": "5.10.8",
     "dotenv": "^16.3.1",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.17.0"
   },
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
-    "test": "jest src"
+    "test": "jest src",
+    "debug": "docker extension dev ui-source mccormsy/gitdocked:latest http://localhost:3000 & docker extension dev debug mccormsy/gitdocked:latest"
   },
   "devDependencies": {
     "@docker/extension-api-client-types": "0.3.4",
